+++ wellregistry/registry/models.py
@@ -1,5 +1,5 @@
 """
-Monitoring Locations Registry ORM object.
+Monitoring Location Registry ORM object.
 """
 
 from django.conf import settings