+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Added Validation for Decimal values in Bulk Upload. These appear as warnings. 
     
 ### Changed
--   Rename/re-title Well Registry to Monitoring Locations Registry
+-   Rename/re-title Well Registry to Monitoring Location Registry
 -   The well_depth field is now allowed to be blank.
 
 
@@ -33,7 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed 
 -   Change text in Well Registry from WL to "water-level" and QW to "water quality"
--   Updated information shown on the monitoring locations list
+-   Updated information shown on the monitoring location list
 
 ## [0.1.0](https://github.com/ACWI-SOGW/well_registry_management/tree/wellregistry-0.1.0) - 2020-11-05